+++ copier/_main.py
@@ -33,7 +33,6 @@ from unicodedata import normalize
 
 from jinja2.exceptions import TemplateError
 from jinja2.loaders import FileSystemLoader
-from jinja2.sandbox import SandboxedEnvironment
 from jinja2.utils import import_string
 from packaging.version import Version
 from pathspec import PathSpec, __version__ as pathspec_version
@@ -45,7 +44,12 @@ from pydantic_core import to_jsonable_python
 from questionary import confirm, unsafe_prompt
 
 from ._deprecation import deprecate_answers_file_template_path
-from ._jinja_ext import IgnoreExtension, YieldExtension, get_yield_context
+from ._jinja_ext import (
+    IgnoreExtension,
+    SandboxedEnvironment,
+    YieldExtension,
+    get_yield_context,
+)
 from ._settings import Settings, SettingsModel, is_trusted_repository
 from ._subproject import Subproject
 from ._template import Task, Template
@@ -83,6 +87,11 @@ from .errors import (
     UserMessageError,
     YieldTagInFileError,
 )
+
+if sys.version_info < (3, 11):
+    from typing_extensions import Self
+else:
+    from typing import Self
 
 _T = TypeVar("_T")
 _P = ParamSpec("_P")
@@ -263,7 +272,7 @@ class Worker:
     answers: AnswersMap = field(default_factory=AnswersMap, init=False)
     _cleanup_hooks: list[Callable[[], None]] = field(default_factory=list, init=False)
 
-    def __enter__(self) -> Worker:
+    def __enter__(self) -> Self:
         """Allow using worker as a context manager."""
         return self
 
@@ -430,7 +439,9 @@ class Worker:
                 for k, v in {**extra_context, "_copier_operation": operation}.items()
             }
             with local.cwd(working_directory), local.env(**extra_env):
-                process = subprocess.run(task_cmd, shell=use_shell, env=dict(local.env))
+                process = subprocess.run(
+                    task_cmd, shell=use_shell, check=False, env=dict(local.env)
+                )
                 if process.returncode:
                     raise TaskError.from_process(process)
 
@@ -622,7 +633,7 @@ class Worker:
                 try:
                     answer = question.parse_answer(self.answers.last[var_name])
                     question.validate_answer(answer)
-                except Exception:
+                except Exception:  # noqa: BLE001
                     del self.answers.last[var_name]
             # Skip a question when the skip condition is met.
             if not question.get_when():
@@ -1071,7 +1082,7 @@ class Worker:
         context pairs.
         """
         if rendered_parts is None:
-            rendered_parts = tuple()
+            rendered_parts = ()
 
         if not parts:
             rendered_path = Path(*rendered_parts)
@@ -1282,7 +1293,7 @@ class Worker:
                 self._render_template()
             if not self.quiet:
                 # TODO Unify printing tools
-                print("")  # padding space
+                print()  # padding space
             if not self.skip_tasks:
                 with Phase.use(Phase.TASKS):
                     self._execute_tasks(self.template.tasks)
@@ -1293,7 +1304,7 @@ class Worker:
         self._print_message(self.template.message_after_copy)
         if not self.quiet:
             # TODO Unify printing tools
-            print("")  # padding space
+            print()  # padding space
 
     @as_operation("copy")
     def run_recopy(self) -> None: