+++ tests/test_prompt.py
@@ -403,7 +403,7 @@ def test_update_with_new_field_in_new_version_skip_answered(
 def test_when(
     tmp_path_factory: pytest.TempPathFactory,
     spawn: Spawn,
-    question_1: str | int | float | bool,
+    question_1: str | float | bool,
     question_2_when: bool | str,
     asks: bool,
 ) -> None:
@@ -1169,6 +1169,7 @@ def test_interactive_session_required_for_question_prompt(
         stdin=subprocess.PIPE,  # Prevents interactive input
         capture_output=True,
         timeout=spawn_timeout or None,
+        check=False,
     )
     assert process.returncode == 1
     assert (
@@ -1197,6 +1198,7 @@ def test_interactive_session_required_for_overwrite_prompt(
         stdin=subprocess.PIPE,  # Prevents interactive input
         capture_output=True,
         timeout=spawn_timeout or None,
+        check=False,
     )
     assert process.returncode == 1
     assert (